Reorganize pool files.
