+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/ProviderBase/DbConnectionClosed.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using Microsoft.Data.Common;
+using Microsoft.Data.SqlClient.ConnectionPool;
 using System.Data;
 using System.Data.Common;
 using System.Diagnostics;
