+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ConnectionPool/DbConnectionPoolAuthenticationContext.cs
@@ -7,7 +7,7 @@ using System.Diagnostics;
 using System.Runtime.ConstrainedExecution;
 using System.Threading;
 
-namespace Microsoft.Data.ProviderBase
+namespace Microsoft.Data.SqlClient.ConnectionPool
 {
     /// <summary>
     /// Represents the context of an authentication attempt when using the new active directory based authentication mechanisms.
