+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ConnectionPool/DbConnectionPoolProviderInfo.cs
@@ -2,13 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Data.ProviderBase
+namespace Microsoft.Data.SqlClient.ConnectionPool
 {
-    partial class DbConnectionPoolIdentity
+    internal class DbConnectionPoolProviderInfo
     {
-        internal static DbConnectionPoolIdentity GetCurrent()
-        {
-            return GetCurrentManaged();
-        }
     }
 }
