+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/ConnectionPool/DbConnectionPoolState.cs
@@ -2,9 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-namespace Microsoft.Data.ProviderBase
+namespace Microsoft.Data.SqlClient.ConnectionPool
 {
-    internal class DbConnectionPoolProviderInfo
+    internal enum DbConnectionPoolState
     {
+        Initializing,
+        Running,
+        ShuttingDown,
     }
 }
